requirements.txt
discord.py>=2.4.0
openai>=1.75.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
aiohttp>=3.11.0
aiosqlite>=0.20.0
requests>=2.32.0
pydantic-settings>=2.0.0
